Add scaleFactor

+++ ros/src/computing/perception/localization/packages/vmml/include/VMap.h
@@ -267,8 +267,8 @@ protected:
 
 	void updateCovisibilityGraph(const kfid k);
 
-	static std::vector<float> mScaleFactors;
-	static std::vector<float> mLevelSigma2;
+	static std::vector<double> mScaleFactors;
+	static std::vector<double> mLevelSigma2;
 
 };
